Backward compatibility for debug tarball, comments, refactor

+++ scripts/AdjustSites.py
@@ -194,7 +194,9 @@ def makeWebDir(ad):
         os.makedirs(path)
         ## Copy the sandbox to the web directory.
         shutil.copy2(os.path.join(".", "sandbox.tar.gz"), os.path.join(path, "sandbox.tar.gz"))
-        shutil.copy2(os.path.join(".", "debug_files.tar.gz"), os.path.join(path, "debug_files.tar.gz"))
+        ## Copy the debug folder. It might not be available if an older (<3.3.1607) crabclient is used.
+        if os.path.isfile(os.path.join(".", "debug_files.tar.gz")):
+            shutil.copy2(os.path.join(".", "debug_files.tar.gz"), os.path.join(path, "debug_files.tar.gz"))
 
         ## Make all the necessary symbolic links in the web directory.
         sourceLinks = ["debug",
